cpu_has_eflag(): mark output registers as earlyclobbers
We assign values to the output registers before the input operand is
used, so we have to use =&r to make sure the input operand isn't
clobbered.

Note: "ri" is correct; it cannot safely be a memory operand, because
the stack pointer isn't where the compiler would expect it to be.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>